requirements.txt

PyQt6>=6.4
google-api-python-client>=2.100
google-auth-oauthlib>=1.0
google-auth>=2.0
opencv-python>=4.5
requests>=2.28
PyMuPDF>=1.22
psutil>=5.8
rawpy>=0.18
numpy>=1.21
imageio>=2.9
Pillow>=9.0
pillow-heif>=0.12